pyproject.toml
[project]
name = "django-unfold-fobi"
version = "0.1.0"
description = "Django Unfold admin integration for django-fobi"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<4.0"
authors = [
    { name = "Marc Widmer", email = "marc@pbi.io" },
]
keywords = ["django", "fobi", "unfold", "admin", "forms"]
dependencies = [
    "django>=5.0",
    "django-unfold>=0.81.0",
    "django-unfold-modal>=0.1.0",
    "django-fobi>=0.19",
    "django-crispy-forms>=2.0",
    "djangorestframework>=3.14",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/metaforx/django-unfold-fobi"
Repository = "https://github.com/metaforx/django-unfold-fobi"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unfold_fobi"]

[tool.poetry]
name = "django-unfold-fobi"
version = "0.1.0"
description = "Django Unfold admin integration for django-fobi"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = ">=5.0"
django-unfold = ">=0.81.0"
django-fobi = {version = ">=0.19", extras = ["simple"]}
django-crispy-forms = ">=2.0"
djangorestframework = ">=3.14"
django-unfold-modal = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0"
pytest-django = ">=4.8"
pytest-playwright = ">=0.5"
playwright = ">=1.40"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.4"

[tool.pytest.ini_options]
addopts = "--ds=testapp.settings"
console_output_style = "progress"
testpaths = ["tests"]
pythonpath = ["src", "tests/server", "tests"]
django_find_project = false
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E501", "B905", "C901"]
